cfgGen| refactor, separate cfg gen script into utility function
